pyproject.toml
[tool.poetry]
name = "innoter"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.7"
psycopg2-binary = "2.9.5"
python-dotenv = "^1.0.0"
djangorestframework = "^3.14.0"
markdown = "^3.4.1"
django-filter = "^22.1"
pyjwt = "^2.6.0"
boto3 = "^1.26.88"
django-storages = "^1.13.2"
pillow = "^9.4.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
